split and moved email_setup.py to email_setup/
