+++ backend/chat_messages/utility/email.py
@@ -1,6 +1,6 @@
 import logging
 
-from climateconnect_api.utility.email_setup import send_email
+from climateconnect_api.utility.email_setup.email_setup import send_email
 from climateconnect_api.utility.translation import (get_user_lang_code,
                                                     get_user_lang_url)
 from django.conf import settings
