+++ backend/ideas/utility/email.py
@@ -1,6 +1,6 @@
 from climateconnect_api.models.user import UserProfile
 from climateconnect_api.utility.translation import get_user_lang_code, get_user_lang_url
-from climateconnect_api.utility.email_setup import send_email
+from climateconnect_api.utility.email_setup.email_setup import send_email
 from django.conf import settings
 
 def send_idea_join_email(user, joining_user, idea, chat_uuid, notification):
